[fix] everything is an int

+++ include/equations.h
@@ -17,75 +17,75 @@ struct Damage;
 
 /* --- MACROS --- */
 /* -- Unit Combat -- */
-#define Equation_multiplyDamage(dmg, multiplier) ((dmg) > (u8)(UINT8_MAX / (multiplier)) ? UINT8_MAX : (u8)((dmg) * (multiplier)))
+#define Equation_multiplyDamage(dmg, multiplier) ((dmg) > (i32)(UINT8_MAX / (multiplier)) ? UINT8_MAX : (i32)((dmg) * (multiplier)))
 
 /* --- FUNCTION DEFINITIONS --- */
 /* -- Unit Computed Stats -- */
-u8 Equation_Unit_Hit(  int wpn_hit,   int dex,       int luck, int support);
-u8 Equation_Unit_Crit( int wpn_crit,  int dex,       int luck, int support);
-i8 Equation_Unit_Speed(int wpn_wgt,   int agi,       int con,  int str);
-u8 Equation_Unit_Favor(int wpn_favor, int faith,     int support);
-i8 Equation_Unit_Dodge(int wpn_wgt,   int wpn_dodge, int luck, int faith,
-                       int agi,       int str,       int con,  int tile_dodge,
-                       int support);
+i32 Equation_Unit_Hit(  i32 wpn_hit,   i32 dex,       i32 luck, i32 support);
+i32 Equation_Unit_Crit( i32 wpn_crit,  i32 dex,       i32 luck, i32 support);
+i32 Equation_Unit_Speed(i32 wpn_wgt,   i32 agi,       i32 con,  i32 str);
+i32 Equation_Unit_Favor(i32 wpn_favor, i32 faith,     i32 support);
+i32 Equation_Unit_Dodge(i32 wpn_wgt,   i32 wpn_dodge, i32 luck, i32 faith,
+                        i32 agi,       i32 str,       i32 con,  i32 tile_dodge,
+                        i32 support);
 
 /* -- Stat total with bonus and malus -- */
-u8 Stat_Total(int current, int bonus, int malus, int cap);
+i32 Stat_Total(i32 current, i32 bonus, i32 malus, i32 cap);
 
 /* -- Staff Healing -- */
-u8 Equation_Staff_Healing(int item_AP,  int user_mag);
+i32 Equation_Staff_Healing(i32 item_AP,  i32 user_mag);
 
 /* -- Unit Healing -- */
-u8 Equation_Unit_Healshp(int total_hp, int heal_percent);
-u8 Equation_Unit_Healshpvar(size_t argnum, ...);
+i32 Equation_Unit_Healshp(i32 total_hp, i32 heal_percent);
+i32 Equation_Unit_Healshpvar(size_t argnum, ...);
 
 /* -- Unit Rescue -- */
-bool Equation_canCarry(int savior_con, int victim_con);
+bool Equation_canCarry(i32 savior_con, i32 victim_con);
 
 /* -- Unit Regrets -- */
-i8 Equation_Regrets(int kills, int faith);
+i32 Equation_Regrets(i32 kills, i32 faith);
 
 /* -- Combat -- */
-u8    Equation_Combat_Hit(     int hit_a,  int avoid_d);
-u8    Equation_Combat_Crit(    int crit_a, int favor_d);
-u8    Equation_Attack_Damage(  int dmg_a,  int def_d);
-u8    Equation_Combat_Damage(  int att_a,  int def_d, int eff, int critp, bool crit);
-float Equation_Expected_Damage(int hit,    int crit,  int critfactor);
+i32    Equation_Combat_Hit(     i32 hit_a,  i32 avoid_d);
+i32    Equation_Combat_Crit(    i32 crit_a, i32 favor_d);
+i32    Equation_Attack_Damage(  i32 dmg_a,  i32 def_d);
+i32    Equation_Combat_Damage(  i32 att_a,  i32 def_d, i32 eff, i32 critp, bool crit);
+float Equation_Expected_Damage(i32 hit,    i32 crit,  i32 critfactor);
 
 /* -- Compute damage total from other damae types in struct -- */
 void  Equation_Damage_Total(struct Damage *dmg);
 
 /* -- Weapon -- */
-u8 Equation_Weapon_Hit(    int Lwpn_hit,    int Rwpn_hit);
-u8 Equation_Weapon_Wgt(    int Lwpn_wgt,    int Rwpn_wgt);
-u8 Equation_Weapon_Crit(   int Lwpn_crit,   int Rwpn_crit);
-u8 Equation_Weapon_Favor(  int Lwpn_favor,  int Rwpn_favor);
-u8 Equation_Weapon_Dodge(  int Lwpn_dodge,  int Rwpn_dodge);
-u8 Equation_Weapon_Attack( int Lwpn_attack, int Rwpn_attack);
-u8 Equation_Weapon_defense(int Lwpn_def,    int Rwpn_def);
+i32 Equation_Weapon_Hit(    i32 Lwpn_hit,    i32 Rwpn_hit);
+i32 Equation_Weapon_Wgt(    i32 Lwpn_wgt,    i32 Rwpn_wgt);
+i32 Equation_Weapon_Crit(   i32 Lwpn_crit,   i32 Rwpn_crit);
+i32 Equation_Weapon_Favor(  i32 Lwpn_favor,  i32 Rwpn_favor);
+i32 Equation_Weapon_Dodge(  i32 Lwpn_dodge,  i32 Rwpn_dodge);
+i32 Equation_Weapon_Attack( i32 Lwpn_attack, i32 Rwpn_attack);
+i32 Equation_Weapon_defense(i32 Lwpn_def,    i32 Rwpn_def);
 
 /* - argvar - */
-u8 Equation_Weapon_Hitvar(    size_t argnum, ...);
-u8 Equation_Weapon_Wgtvar(    size_t argnum, ...);
-u8 Equation_Weapon_Critvar(   size_t argnum, ...);
-u8 Equation_Weapon_Favorvar(  size_t argnum, ...);
-u8 Equation_Weapon_Dodgevar(  size_t argnum, ...);
-u8 Equation_Weapon_Attackvar( size_t argnum, ...);
-u8 Equation_Weapon_Defensevar(size_t argnum, ...);
+i32 Equation_Weapon_Hitvar(    size_t argnum, ...);
+i32 Equation_Weapon_Wgtvar(    size_t argnum, ...);
+i32 Equation_Weapon_Critvar(   size_t argnum, ...);
+i32 Equation_Weapon_Favorvar(  size_t argnum, ...);
+i32 Equation_Weapon_Dodgevar(  size_t argnum, ...);
+i32 Equation_Weapon_Attackvar( size_t argnum, ...);
+i32 Equation_Weapon_Defensevar(size_t argnum, ...);
 
 /* -- Agony -- */
-u8 Equation_Agony_Turns(        int str, int def, int con);
-u8 Equation_Agony_PercentonCrit(int luck, int con);
+i32 Equation_Agony_Turns(        i32 str, i32 def, i32 con);
+i32 Equation_Agony_PercentonCrit(i32 luck, i32 con);
 
 /* -- AI equations -- */
-i8 Equation_AI_Rating_Stats(   int   stat);
-i8 Equation_AI_Rating_Damage(  int   dmg);
-i8 Equation_AI_Rating_HPprop(  float hp_prop);
-i8 Equation_AI_Rating_hitRate( int   hit_rate);
-i8 Equation_AI_Rating_critRate(int   crit_rate);
+i32 Equation_AI_Rating_Stats(   i32   stat);
+i32 Equation_AI_Rating_Damage(  i32   dmg);
+i32 Equation_AI_Rating_HPprop(  float hp_prop);
+i32 Equation_AI_Rating_hitRate( i32   hit_rate);
+i32 Equation_AI_Rating_critRate(i32   crit_rate);
 
-i8 Equation_AI_Rating_Equipment(   int crit);
-i8 Equation_AI_Rating_pushGradient(i32 gradient);
-i8 Equation_AI_Rating_pullGradient(i32 gradient);
+i32 Equation_AI_Rating_Equipment(   i32 crit);
+i32 Equation_AI_Rating_pushGradient(i32 gradient);
+i32 Equation_AI_Rating_pullGradient(i32 gradient);
 
 #endif /* EQUATIONS_H */
